申告を実測に合わせ、音声を bfloat16 に揃える
VRAM の申告が実態の 1.5〜4 倍だった。多めに言うと、LLM が載っている間は空きが
あっても弾かれる。実測（2026-09-05、gfx1201 / R9700、GPU 占有）:

  speech.asr.transcribe   5GB 申告 →  1.84 GiB 実測  →  3GB へ
  speech.tts.synthesize   8GB 申告 →  2.49 GiB 実測  →  4GB へ
  music.*                18GB 申告 → 11.78 GiB 実測  → 13GB へ

LLM 21.42 GiB 常駐時の残りは 10.44 GiB なので、実測値なら ASR も TTS も共存できる。
以前の申告では全部弾かれていた。

whisper が float16 のままだったので bfloat16 に揃えた。float16 は指数部が狭く、
logits が溢れると確率に NaN が入る。その NaN を torch.multinomial が踏むと ROCm
では assert カーネルが HIP 719（unspecified launch failure）で落ちる。単体で
再現済みで、正常な確率は語彙 20 万でも通り、NaN を混ぜると落ちる。ASR は貪欲
デコードなので今は multinomial を通らないが、同じ桁溢れの上に乗っている。

実測でも bfloat16 の方が速い（ASR 5 秒の音声で 4.29 → 3.66 秒、VRAM は同じ）。
TTS 側は元から bfloat16 で、実機で 6.80 秒の音声を正常に生成できている。

+++ worker_packs/whisper/live_worker.py
@@ -41,7 +41,12 @@ def asr_pipe(model_id: str):
     if cached is not None:
         return cached
     device = 0 if torch.cuda.is_available() else -1
-    dtype = torch.float16 if device >= 0 else torch.float32
+    # bfloat16 にする。float16 は指数部が狭く、logits が溢れると確率に NaN が
+    # 入る。その NaN を torch.multinomial が踏むと、ROCm では assert カーネルが
+    # HIP 719（unspecified launch failure）で落ちる（2026-09-05 に単体再現）。
+    # ASR は貪欲デコードなので今は multinomial を通らないが、同じ桁溢れの上に
+    # 乗っている。TTS 側は既に bfloat16 で、そちらは実機で正常に生成できている。
+    dtype = torch.bfloat16 if device >= 0 else torch.float32
     model = AutoModelForSpeechSeq2Seq.from_pretrained(
         model_id,
         torch_dtype=dtype,
